Changed notification logic to a hook
